services/connection-service/requirements.txt
fastapi
uvicorn
websockets
kafka-python
